Fix build break
(cherry picked from commit 3b9acf6afdb2fc6aec96a0d3816332cdb060bd39)

+++ gcc/reg-stack.c
@@ -2750,7 +2750,7 @@ convert_regs_entry (void)
     {
       rtx x = DECL_INCOMING_RTL (parm);
       if (STACK_REG_P (x))
-	incoming_arg[REGNO (x) - FIRST_STACK_REG] = x;
+	incoming_arg[REGNO (x) - FIRST_STACK_REG] = (int) x;
       parm = TREE_CHAIN (parm);
     }
